Step 3

+++ src/main/java/ru/hogwarts/school/service/AvatarService.java
@@ -7,47 +7,55 @@ import org.springframework.web.multipart.MultipartFile;
 import ru.hogwarts.school.model.Avatar;
 import ru.hogwarts.school.model.Student;
 import ru.hogwarts.school.repository.AvatarRepository;
+import ru.hogwarts.school.repository.StudentRepository;
 
 import java.io.*;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.util.Optional;
+
 import static java.nio.file.StandardOpenOption.CREATE_NEW;
 
 @Service
 @Transactional
 public class AvatarService {
-    private final StudentService studentService;
+    private final StudentRepository studentRepository;
     private final AvatarRepository avatarRepository;
 
     @Value("${path.to.avatars.folder}")
     private String avatarsDir;
 
-    public AvatarService(StudentService studentService, AvatarRepository avatarRepository) {
-        this.studentService = studentService;
+    public AvatarService(StudentRepository studentRepository, AvatarRepository avatarRepository) {
+        this.studentRepository = studentRepository;
         this.avatarRepository = avatarRepository;
     }
 
     public void uploadAvatar(Long studentId, MultipartFile avatarFile) throws IOException {
-        Student student = studentService.readStudent(studentId);
-        Path filePath = Path.of(avatarsDir, student + "." + getExtensions(avatarFile.getOriginalFilename()));
-        Files.createDirectories(filePath.getParent());
-        Files.deleteIfExists(filePath);
-        try (
-                InputStream is = avatarFile.getInputStream();
-                OutputStream os = Files.newOutputStream(filePath, CREATE_NEW);
-                BufferedInputStream bis = new BufferedInputStream(is, 1024);
-                BufferedOutputStream bos = new BufferedOutputStream(os, 1024);
-        ) {
-            bis.transferTo(bos);
+        Optional<Student> optionalStudent = studentRepository.findById(studentId);
+        if (optionalStudent.isPresent()) {
+            Student student = optionalStudent.get();
+            Path filePath = Path.of(avatarsDir,
+                    student.getId() + "." + getExtensions(avatarFile.getOriginalFilename()));
+            Files.createDirectories(filePath.getParent());
+            Files.deleteIfExists(filePath);
+            try (
+                    InputStream is = avatarFile.getInputStream();
+                    OutputStream os = Files.newOutputStream(filePath, CREATE_NEW);
+                    BufferedInputStream bis = new BufferedInputStream(is, 1024);
+                    BufferedOutputStream bos = new BufferedOutputStream(os, 1024);
+            ) {
+                bis.transferTo(bos);
+            }
+            Avatar avatar = readAvatarByStudent(studentId);
+            avatar.setStudent(student);
+            avatar.setFilePath(filePath.toString());
+            avatar.setFileSize(avatarFile.getSize());
+            avatar.setMediaType(avatarFile.getContentType());
+            avatar.setData(avatarFile.getBytes());
+            avatarRepository.save(avatar);
         }
-        Avatar avatar = readAvatarByStudent(studentId);
-        avatar.setStudent(student);
-        avatar.setFilePath(filePath.toString());
-        avatar.setFileSize(avatarFile.getSize());
-        avatar.setMediaType(avatarFile.getContentType());
-        avatar.setData(avatarFile.getBytes());
-        avatarRepository.save(avatar);
     }
+
     private String getExtensions(String fileName) {
         return fileName.substring(fileName.lastIndexOf(".") + 1);
     }
@@ -55,8 +63,8 @@ public class AvatarService {
     public Avatar readAvatar(Long id) {
         return avatarRepository.findById(id).orElse(null);
     }
-    public Avatar readAvatarByStudent(Long id) {
-        return avatarRepository.findByStudentId(id).orElse(new Avatar());
+    public Avatar readAvatarByStudent(Long studentId) {
+        return avatarRepository.findByStudentId(studentId).orElse(new Avatar());
     }
 
 }
